split truncation apart into separate files
