+++ libs/core/routing/src/uri/public_api.ts
@@ -1,2 +1,2 @@
-export * from './truncation';
+export * from './truncation/public_api';
 export { DaffRoutingUriNormalizer } from './normalizer.service';
