+++ libs/core/routing/src/uri/truncation/leading-path-segments.ts
@@ -1,0 +1,6 @@
+/**
+ * Truncates leading path segments from a URI.
+ * e.g. /foo/bar/baz.html -> baz.html
+ */
+export const daffUriTruncateLeadingPathSegments = (uri: string): string =>
+  uri.split('/').pop();
