+++ libs/core/routing/src/uri/truncation/query-fragment.ts
@@ -1,0 +1,6 @@
+/**
+ * Truncates query parameters and fragments from a URI.
+ * e.g. /foo/bar/baz.html?thing=test#id -> /foo/bar/baz.html
+ */
+export const daffUriTruncateQueryFragment = (uri: string): string =>
+  (new URL(`http://fakedomain.com${uri}`)).pathname;
